backend/requirements.txt
# FastAPI - versioni compatibili con Python 3.12+
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Config - versioni aggiornate per evitare compilazione Rust
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17

# HTTP Client
httpx==0.27.2

# Redis
redis==5.2.0

# Utils
python-slugify==8.0.4
bleach==6.2.0

# Storage (S3/R2 compatible)
boto3==1.35.0
aiobotocore==2.15.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
